Remove unnecessary synchronized
The only issue is that if we should wait unil the function succeed to return.
There are 2 possible way this method is invoked. If not success in constructor method, only current thread might access this method, otherwise, this method will only be access by Finalizer thread. And we are thread safe already.